requirements.txt
fastapi
uvicorn
psutil
pillow
